requirements.txt
streamlit>=1.28.0
tensorflow>=2.13.0
numpy>=1.23.0
Pillow>=9.0.0
